requirements.txt
fastapi==0.95.0
gunicorn==20.1.0
h5py==3.8.0
keras==2.12.0
numpy==1.23.5
opencv-python==4.7.0.72
Pillow==9.4.0
python-multipart==0.0.6
scipy==1.10.1
tensorflow==2.12.0
uvicorn==0.21.1

